Faster uv hooks

+++ modules/projects/tools/uv/uv-shell-hook.sh
@@ -3,28 +3,36 @@
 uvShellHook() {
   echo "Executing uvShellHook"
 
+  export VIRTUAL_ENV_DISABLE_PROMPT=1
+
+  local -a venvPackages
+  runHook venvPackages
+
   # Check if we need to recreate the virtual environment
-  if [ "$(readlink .venv/nix-env)" != "${NIX_GCROOT}" ]; then
+  if [ "$(readlink .venv/nix-env)" == "${NIX_GCROOT}" ] && [ "${PYTHONPATH}" == "$(cat .venv/python-path)" ]; then
+    source .venv/bin/activate
+  else
     # Remove the old virtual environment if it exists
     rm -rf .venv
+    echo "Recreating the python virtual environment"
+    uv venv
+    source .venv/bin/activate
+
+    # Configure the virtual environment to automatically pick up the Nix PYTHONPATH even if it is not activated
+    # inside a nix develop shell (for VSCode & PyCharm)
+    _SITE_PACKAGES_DIR=$(python -c "import site; print(site.getsitepackages()[0])")
+    echo "import _nix_env" >"$_SITE_PACKAGES_DIR/_nix_env.pth"
+    echo "import site" >"$_SITE_PACKAGES_DIR/_nix_env.py"
+    for PYTHON_PATH_COMPONENT in $(echo "$PYTHONPATH" | tr ':' $'\n'); do
+      echo "site.addsitedir(\"$PYTHON_PATH_COMPONENT\")" >>"$_SITE_PACKAGES_DIR/_nix_env.py"
+    done
+
+    uv pip install --no-deps --offline --no-cache --no-build-isolation "${venvPackages[@]}"
+
+    ln -fs "${NIX_GCROOT}" .venv/nix-env
+    echo "${PYTHONPATH}" >.venv/python-path
   fi
 
-  uv venv --allow-existing
-  export VIRTUAL_ENV_DISABLE_PROMPT=1
-  source .venv/bin/activate
-
-  # Configure the virtual environment to automatically pick up the Nix PYTHONPATH even if it is not activated
-  # inside a nix develop shell (for VSCode & PyCharm)
-  _SITE_PACKAGES_DIR=$(python -c "import site; print(site.getsitepackages()[0])")
-  echo "import _nix_env" >"$_SITE_PACKAGES_DIR/_nix_env.pth"
-  echo "import site" >"$_SITE_PACKAGES_DIR/_nix_env.py"
-  for PYTHON_PATH_COMPONENT in $(echo "$PYTHONPATH" | tr ':' $'\n'); do
-    echo "site.addsitedir(\"$PYTHON_PATH_COMPONENT\")" >>"$_SITE_PACKAGES_DIR/_nix_env.py"
-  done
-
-  ln -fs "${NIX_GCROOT}" .venv/nix-env
-
-  runHook postVenvCreation
   echo "Finished executing uvShellHook"
 }
 
@@ -44,18 +52,18 @@ findconfig() {
   fi
 }
 
-@name@PostVenvCreationHook() {
-  echo "Executing @name@PostVenvCreationHook"
+@name@VenvPackagesHook() {
+  echo "Executing @name@VenvPackagesHook"
   local projectRoot
   projectRoot="$(dirname "$(findconfig flake.nix)")"
   if test -f "${projectRoot}/@relativePathToRoot@/pyproject.toml"; then
-    uv pip install -e "${projectRoot}/@relativePathToRoot@" --no-deps --offline --no-cache --no-build-isolation
+    venvPackages+=(--editable "${projectRoot}/@relativePathToRoot@")
     addToSearchPath PYTHONPATH "${projectRoot}/@relativePathToRoot@/src"
   fi
 }
 
-if [ -z "${postVenvCreationHooks:-}" ]; then
+if [ -z "${venvPackagesHooks:-}" ]; then
   preShellHooks+=(uvShellHook)
 fi
 
-postVenvCreationHooks+=(@name@PostVenvCreationHook)
+venvPackagesHooks+=(@name@VenvPackagesHook)
